fix: add the Local widget in the navbar
