+++ sepal_ui/sepalwidgets/app.py
@@ -1,15 +1,18 @@
 from functools import partial
 from datetime import datetime
+from pathlib import Path
 
 import ipyvuetify as v
 from deprecated.sphinx import versionadded
+import pandas as pd
 
 from sepal_ui.sepalwidgets.sepalwidget import SepalWidget
 from sepal_ui import color
 from sepal_ui.frontend.styles import sepal_main, sepal_darker
 from sepal_ui.frontend import js
+from sepal_ui.scripts import utils as su
 
-__all__ = ["AppBar", "DrawerItem", "NavDrawer", "Footer", "App"]
+__all__ = ["AppBar", "DrawerItem", "NavDrawer", "Footer", "App", "localeSelect"]
 
 
 class AppBar(v.AppBar, SepalWidget):
@@ -36,12 +39,14 @@ class AppBar(v.AppBar, SepalWidget):
 
         self.title = v.ToolbarTitle(children=[title])
 
+        self.locale = localeSelect()
+
         # set the default parameters
         kwargs["color"] = kwargs.pop("color", sepal_main)
         kwargs["class_"] = kwargs.pop("class_", "white--text")
         kwargs["dense"] = kwargs.pop("dense", True)
         kwargs["app"] = True
-        kwargs["children"] = [self.toggle_button, self.title]
+        kwargs["children"] = [self.toggle_button, self.title, v.Spacer(), self.locale]
 
         super().__init__(**kwargs)
 
@@ -380,3 +385,90 @@ class App(v.App, SepalWidget):
         self.content.children = [alert] + self.content.children.copy()
 
         return self
+
+
+class localeSelect(v.Menu, SepalWidget):
+    """
+    An language selector for sepal-ui based application.
+    it displays the currently requested language (not the one used by the translator).
+    When value is changed, the sepal-ui config file is updated
+    """
+
+    COUNTRIES = pd.read_csv(Path(__file__).parents[1] / "scripts" / "locale.csv")
+    "pandas.DataFrame: the country list as a df. columns [code, name, flag]"
+
+    FLAG = "https://flagcdn.com/{}.svg"
+    "str: the url of the svg flag images"
+
+    ATTR = {"src": "https://flagcdn.com/gb.svg", "width": "30", "alt": "en-UK"}
+    "dict: the default flag parameter, default to english"
+
+    def __init__(self):
+
+        self.btn = v.Btn(
+            small=True,
+            v_model=False,
+            v_on="x.on",
+            children=[v.Html(tag="img", attributes=self.ATTR, class_="mr-1"), "en-UK"],
+        )
+
+        self.language_list = v.List(
+            dense=True,
+            flat=True,
+            color="grey darken-3",
+            v_model=True,
+            max_height="300px",
+            style_="overflow: auto; border-radius: 0 0 0 0;",
+            children=[v.ListItemGroup(children=self._get_country_items(), v_model="")],
+        )
+
+        super().__init__(
+            children=[self.language_list],
+            v_model=False,
+            close_on_content_click=True,
+            v_slots=[{"name": "activator", "variable": "x", "children": self.btn}],
+        )
+
+        # add js behaviour
+        self.language_list.children[0].observe(self._on_locale_select, "v_model")
+
+    def _get_country_items(self):
+        """get the list of countries in as a list of listItem"""
+
+        country_list = []
+        for r in self.COUNTRIES.itertuples(index=False):
+
+            attr = {**self.ATTR, "src": self.FLAG.format(r.flag), "alt": r.name}
+
+            children = [
+                v.ListItemAction(children=[v.Html(tag="img", attributes=attr)]),
+                v.ListItemContent(children=[v.ListItemTitle(children=[r.name])]),
+                v.ListItemActionText(children=[r.code]),
+            ]
+
+            country_list.append(v.ListItem(value=r.code, children=children))
+
+        return country_list
+
+    def _on_locale_select(self, change):
+        """
+        adapt the application to the newly selected language
+
+        Display the new flag and country code on the widget btn
+        change the value in the config file
+        """
+
+        # get the line in the locale dataframe
+        loc = self.COUNTRIES[self.COUNTRIES.code == change["new"]].squeeze()
+
+        # change the btn attributes
+        attr = {**self.ATTR, "src": self.FLAG.format(loc.flag), "alt": loc.name}
+        self.btn.children = [
+            v.Html(tag="img", attributes=attr, class_="mr-1"),
+            loc.code,
+        ]
+
+        # change the paramater file
+        su.set_config_locale(loc.code)
+
+        return
